gee_learn_from_geektutu/day7-panic-recover/gee/router_test.go
package gee

import (
	"fmt"
	"reflect"
	"testing"
)

func newTestRouter() *router {
	r := newRouter()
	r.addRoute("GET", "/", nil)
	r.addRoute("GET", "/hello/:name", nil)
	r.addRoute("GET", "/hello/b/c", nil)
	r.addRoute("GET", "/hi/:name", nil)
	r.addRoute("GET", "/assets/*filepath", nil)
	return r
}

// go test 必须以Test开头，后缀必须以大写字母开头
func TestParsePattern(t *testing.T) {
	ok := reflect.DeepEqual(parsePattern("/p/:name"), []string{"p", ":name"})
	ok = ok && reflect.DeepEqual(parsePattern("/p/*"), []string{"p", "*"})
	ok = ok && reflect.DeepEqual(parsePattern("/p/*name/*"), []string{"p", "*name"})
	if !ok {
		t.Fatal("test parsePattern failed")
	}
}

func TestGetRoute(t *testing.T) {
	r := newTestRouter()
	n, ps := r.getRoute("GET", "/hello/yitong")

	if n == nil {
		t.Fatal("n should not be nil")
	}

	if n.pattern != "/hello/:name" {
		t.Fatal("should match /hello/:name")
	}

	if ps["name"] != "yitong" {
		t.Fatal("name should be equal to 'yitong'")
	}

	fmt.Printf("matched path: %s, params['name']: %s\n", n.pattern, ps["name"])
}

func TestGetRoute2(t *testing.T) {
	r := newTestRouter()
	n1, ps1 := r.getRoute("GET", "/assets/file.txt")

	if n1.pattern != "/assets/*filepath" || ps1["filepath"] != "file.txt" {
		t.Fatal("pattern should be /assets/*filepath & filepath should be file.txt")
	}

	n2, ps2 := r.getRoute("GET", "/assets/css/test.css")
	if n2.pattern != "/assets/*filepath" || ps2["filepath"] != "css/test.css" {
		t.Fatal("pattern should be /assets/*filepath & filepath shoud be css/test.css")
	}
}

func TestGetRoutes(t *testing.T) {
	r := newTestRouter()
	nodes := r.getRoutes("GET")
	for i, n := range nodes {
		fmt.Println(i+1, n)
	}

	if len(nodes) != 5 {
		t.Fatal("the number of routes should be 5")
	}
}
